+++ src/main/kotlin/component/fragment/src/createFragmentRecipe.kt
@@ -4,7 +4,11 @@ import com.android.tools.idea.wizard.template.ModuleTemplateData
 import com.android.tools.idea.wizard.template.RecipeExecutor
 import component.activity.res.layout.createActivityXml
 import component.fragment.src.app_package.createFragmentKt
+import component.router.*
+import component.util.appendToJavaFile
+import component.util.appendToSecondLineFromTheBottom
 import component.viewmodel.createViewModelKt
+import java.io.File
 
 /**
  * Author：Stevie.Chen Time：2023/8/30
@@ -47,5 +51,78 @@ fun RecipeExecutor.createFragmentRecipe(
             createViewModelKt(packageName, fragmentClass),
             srcOut.resolve("viewmodel/${fragmentClass}ViewModel.${ktOrJavaExt}")
         )
+    }
+    if (isRouter) {
+        val routerRootPath = "biz_common/src/main/java/com/zondy/biz/common/router"
+        // 生成Router
+        val routerFile = File(
+            projectData.rootDir,
+            "$routerRootPath/ARoutePath.${ktOrJavaExt}"
+        )
+        if (routerFile.exists()) {
+            routerFile.appendToJavaFile(
+                createRouterKt(
+                    packageName,
+                    fragmentClass,
+                    false
+                ),
+                getRouterDeclare(fragmentClass)
+            )
+        } else {
+            save(
+                createRouterKt(packageName, fragmentClass, true),
+                routerFile
+            )
+        }
+        // 生成RouterService
+        val routerServiceFile = File(
+            projectData.rootDir,
+            "$routerRootPath/service/RouterService.${ktOrJavaExt}"
+        )
+        if (routerServiceFile.exists()) {
+            routerServiceFile.appendToJavaFile(
+                createRouterServiceKt(
+                    fragmentClass,
+                    isCreate = false,
+                    isActivity = false
+                ),
+                getRouterServiceDeclare(fragmentClass, isActivity = false)
+            )
+        } else {
+            save(
+                createRouterServiceKt(
+                    fragmentClass,
+                    isCreate = true,
+                    isActivity = false
+                ),
+                routerServiceFile
+            )
+        }
+        // 生成RouterServiceImpl
+        val routerServiceImplFile = File(
+            projectData.rootDir,
+            "$routerRootPath/service/RouterServiceIml.${ktOrJavaExt}"
+        )
+        if (routerServiceImplFile.exists()) {
+            routerServiceImplFile.appendToJavaFile(
+                createRouterServiceImplKt(
+                    packageName,
+                    fragmentClass,
+                    isCreate = false,
+                    isActivity = false
+                ),
+                getRouterServiceImplDeclare(fragmentClass, isActivity = false)
+            )
+        } else {
+            save(
+                createRouterServiceImplKt(
+                    packageName,
+                    fragmentClass,
+                    isCreate = true,
+                    isActivity = false
+                ),
+                routerServiceImplFile
+            )
+        }
     }
 }